Move makeRendererSafe method to a util file in grid/impl
