+++ format/FormatUtils.js
@@ -44,18 +44,3 @@ export function asElementDeprecationWarning(opts) {
         msg: 'Formatters return elements by default. You can use `asHtml` to return a HTML string'
     });
 }
-
-/**
- * Used to ensure grid renderers cannot bring down an application
- */
-export function makeRendererSafe(fn) {
-    if (!isFunction(fn)) return fn;
-    return (...args) => {
-        try {
-            return fn(...args)
-        } catch (e) {
-            console.warn('A grid renderer has thrown an error.', e);
-            return '#ERROR'
-        }
-    }
-}
